Set and get base URL and port for cromwell server

The Cromwell server presents a RESTFul API. The base URL is of the form: `http://EXAMPLE.COM:PORT`. The current approach to changing that url is to set an option, `cromwellBase` to a valid URL (without trailing slash). This URL will then be used throughout the `cRomwell` package. If no option is set, the server is assumed to be running at `http://localhost:8000`.

cromwellBase()

setCromwellBase(base_url = NULL)

Arguments

base_url
character(1) specifying the url and port of the cromwell server

Value

A character(1) base url (default http://localhost:8000)

Examples

cromwellBase()

#> [1] "http://example.com:8111"

# set a bogus host
setCromwellBase('http://example.com:8111')
cromwellBase()

#> [1] "http://example.com:8111"

# and set back to NULL to get the default behavior
setCromwellBase()

#> Error: is.character(base_url) is not TRUE

cromwellBase()

#> [1] "http://example.com:8111"