Get the info about cromwell workflows

Each of the following terms can be specified one or more times. Simply create a named list or named character vector.

cromwellQuery(query = NULL, ...)

Arguments

query
terms about the workflow

...
passed directly to httr `GET` (for including `timeouts`, `handles`, etc.)

Value

a data.frame of query results

Details

name

The name of a job; may be specified more than once

status

one of Succeeded, Failed, Running

id

an id of a cromwell job

start

a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset

end

a timestamp of the form "2015-11-01T07:45:52.000-05:00", including mandatory offset

page

if paging is used, what page to select

pagesize

if paging is used, how many records per page

Examples

## Not run: ------------------------------------
# res = cromwellQuery(terms=c(status='Succeeded',name='taskName'))
# head(res)
## ---------------------------------------------