return inputs from wdltool

This little function just runs the command-line wdltool inputs on a wdl file and returns a data.frame of input names and input types.

wdltoolInputs(wdlfile, wdltoolJar = Sys.getenv("WDLTOOL_JAR", "wdltool.jar"))

Arguments

wdlfile

The filename of a WDL file from which to get inputs.

wdltoolJar

The filename of the wdltool jar file; if not specified, the environment variable WDLTOOL_JAR is consulted.

Value

a 2-column data.frame with the input name and the input type as the two columns.

Examples

version = '0.8'
tmpfile = file.path(tempdir(),'wdltool.jar')
fp = getWdltoolJar(wdltool_version = version,
                   destfile = tmpfile)

#> WDLTOOL_JAR environment variable set to
#> /var/folders/21/b_rp6qyj1_b1j5cp8qby0tnr0000gn/T//RtmpKPNq6f/wdltool.jar

wdltoolInputs(system.file(package='cRomwell','wdl/fastqc_task.wdl'))

#>                      name type
#> 1 r_fqa_wf.r_fastqa.fastq File